Fixed displayColor primvar population when defined but empty
Previously, if no displayColor/displayOpacity primvars were specified for a prim, we defaulted to grey color with constant interp mode. However, it is perfectly legal to define an empty displayColor/displayOpacity primvar (i.e., with zero values) and populate it over time, as is done in vfx assets authored in houdini.
By marking it as constant, we overrule the user specification (bad) and spew incorrect warnings when the primvar is populated with data as a result of the value cache expecting a single value rather than an array.

The fix ensures that empty displayColor and/or displayOpacity primvar(s) are handled correctly -- i.e., the interp mode isn't changed and the default value isn't added.

(Internal change: 1754831)