Fixed recursive plugInfo search on Windows.
The recursive search for plugInfo.json files uses TfStringCatPaths()
and a regex.  TfStringCatPaths() calls TfNormPath() which, on Windows,
will lowercase the drive letter.  On the other hand, the regex does a
case sensitive drive letter match.  So if we pass a search directory
with an uppercase drive letter then we'll never match because the
regex will have uppercase and all the generated paths will have
lowercase.  Now we use TfNormPath() before creating the regex so it
will also have lowercase.

(Internal change: 1755253)