Added support for static and monolithic libraries.
The cmake files now have the concept of "core" libraries and
options for a monolithic build.  In a monolithic build, enabled
by setting PXR_BUILD_MONOLITHIC to ON, the core libraries are
accumulated into a single static (archive) library, usd_m.

If PXR_BUILD_MONOLITHIC is OFF (the default) and BUILD_SHARED_LIBS
is ON (the default) then we build a shared library for each core
library just as we used to.  If PXR_BUILD_MONOLITHIC is OFF and
BUILD_SHARED_LIBS is OFF we build a static library for each core
library.  In this case we're not building a shared library so we
can't build Python modules or anything that uses Python modules.

If PXR_BUILD_MONOLITHIC is ON then BUILD_SHARED_LIBS is ignored.
We build a single archive library, usd_m, and from that a single
monolithic shared library, usd_ms.  We then build the Python
modules and plugins using usd_ms.

However, if PXR_MONOLITHIC_IMPORT is set then we have a two phase
build.  In phase 1 we build the usd_m library and in phase 2 we
use a client-provided shared library to complete the build.  In
phase 2, the file pointed to by PXR_MONOLITHIC_IMPORT must have
valid cmake contents that creates a SHARED IMPORT library named
usd_ms.  During phase 1 the file doesn't need to exist but if it
does, it should be empty.  Instead of being empty, it may have a
SHARED IMPORT usd_ms library from a previous build;  this will be
ignored.

The two phase build has the following steps:  1) configure with
cmake, with PXR_BUILD_MONOLITHIC=ON and PXR_MONOLITHIC_IMPORT
set to the location where the usd_ms export file will be;  2)
build with cmake using the target 'monolithic';  3) build the
shared library, including the whole usd_m archive;  4) modify
the PXR_MONOLITHIC_IMPORT export file if necessary;  5) finish
the USD build using the default or 'install' target.

Phase 1 will generate but not install the usd_m library and two
files in the top of the build tree: usd-targets-<CONFIG>.cmake
that can be used to import the usd_m library in another cmake
build, and usd-imports-<CONFIG>.cmake that provides boilerplate
for the client's own export file.  The client may choose to use
usd-imports-<CONFIG>.cmake with some edits or have cmake generate
an entirely new file.

Each platform has a different way to link the entire contents of
an archive library.  On Linux it's "--whole-archive <lib>
--no-whole-archive"; on Darwin it's "-force_load <lib>"; and on
Windows it's "/WHOLEARCHIVE:<lib>".  If the whole archive isn't
included then the shared library will not include global objects
that are not referenced outside their translation unit even if
they have side effects.  USD relies on these extensively so the
resulting shared library will not work correctly.

The target 'python' was added since it was simple to do so.
This target depends on anything that is or uses a Python module.
It's automatically added to the ALL target to build everything
outside the core libraries when BUILD_SHARED_LIBS is ON or the
usd_ms target exists (which is when PXR_BUILD_MONOLITHIC is ON
and PXR_MONOLITHIC_IMPORT isn't set or it is set and it points
to a file with an usd_ms SHARED IMPORT library).

Some tests were adjusted.  TfDebugTestEnv, TfDebugTestEnvList and
testUsdGeomNoPlugLoad are disabled in a monolithic build.  The
first two are disabled because TF_DEBUG symbols in a monolithic
are reported from every core library and the Tf test shouldn't
depend on downstream libraries.  testUsdGeomNoPlugLoad is disabled
because it requires plugins to not be loaded but most "plugins"
are core libraries and thus part of the monolithic library, which
must be loaded.

(Internal change: 1756108)