Fixed handling of files on Windows in testSdfParsing.
On Windows we can't have a file open and also unlink it.  This
change passes delete=False to NamedTemporaryFile() to ensure that
we have manual control over file deletion (which was the clear
expectation) and we immediately close the file after creation
so we can create/close/erase it later, which we do once per test
file.

(Internal change: 1756110)