Changed numbers that cause rounding differences between platforms.
The numbers 3.14159 and 10.403 read as 3.140625 and 10.40625,
respectively, but how they're written depends on the platform and,
possibly, the rounding mode.  On Linux and Darwin 3.140625 is
written as 3.14062 unless the rounding mode is FE_UPWARD, in which
case it's written as 3.14063.  (Similarly for 10.40625.)  On
Windows 3.140625 and 10.40625 are always written as 3.14063 and
10.4063, respectively, causing a diff failure.

This change simply changes the values to 3.14257 and 10.415.
These will round trip correctly.

(Internal change: 1756112)