Handle signed enum bitfields on Windows in pcp.
PcpPrimIndex_Graph::_Node::_SmallInts has a bitfield for a
PcpArcType value, allocating 4 bits.  We need all four bits
but on Windows enums are stored as signed values so we
actually need an additional sign bit to get the desired
behavior.  We were using 30 bits out of 32 so this doesn't
increase the size of the structure.

Other solutions are to cast the value or use an enum class
with an unsigned base type.  Both require larger fixes to the
code.  The former requires casting everywhere the value is
used while the latter requires scoping the enumerators.

This change was necessary because specializations were not
appearing in the right place in the opinion stack on Windows
because of the signed comparison (they were treated as negative
and so appeared first instead of last).

(Internal change: 1756113)