Pre-Sync Dirty Bit Propagation (Backed out again).
Back out changelist 1751291.

The dirty bit propagation change intro another bug in Maya.  In this case the usage pattern is that Maya is introducing a new root path to the collection, which is causing the repr to not be initialized.  We have a fix for this issue, but it could possibly introduce further performance regression.  Fixes for these performance issues are already in flight, but going to defer making these changes until the full set of fixes and optimization are in place, so they can all be included in the same release.

(Internal change: 1756508)