testUsdFileFormats fixes for Windows.
The test assumed /tmp was a valid directory.  We now use tempfile
to generate a temporary file.

We must ensure the layer for the temporary file is released so it
will close the file before we try to delete the file.  On Windows
you can't delete an open file.

(Internal change: 1757275)