Fixed testUsdStageLoadUnload on Windows.
We need to be sure that we don't have a file open if we're going
to try to move or remove it.  However in this case we want to
open a crate file, modify its contents then use Reload() to load
those new contents.  This might be possible on Windows but we
currently get access denied or invalid argument when we try to
replace the contents.  So this change disables the stage layer
reload test for crate files on Windows.

(Internal change: 1757283)