Fix memory corruption bug in Pcp
Pcp_ParallelIndexer held a reference to a PcpLayerStackPtr
that had been converted from a PcpLayerStackRefPtr when
it was constructed. This temporary PcpLayerStackPtr
expired when the constructor completed, which led to
problems when the reference was accessed later on.

This was a long-standing bug and did not appear to
cause issues except on Xcode 8.3.2.

(Internal change: 1757406)