Changed usddiff to use Python's difflib.
Previously usddiff used external tools to diff ('diff' on Linux
and Mac, 'fc' on Windows).  This worked but the output was
different.

Now we use difflib.unified_diff() to get the same result on all
platforms.  The baselines were appopriately updated.  The test
testDiffToolSelectionSingleEnvVar is no longer relevant since
you can't select the external diff tool anymore.  The test
testDiffToolSelectionMultiEnvVar seemed identical to the test
testDiffToolSelectionDefault.  Both were removed.

usddiff still respects the USD_DIFF and DIFF environment
variables.  If either is set we'll use the tool specified and
if the named tool doesn't exist we'll error out.  If neither
is set we fall back to the unified diff.

(Internal change: 1757506)