Fixes for crate files on Windows.
fopen_s() is more restrictive than fopen() when the files are
already open.  Switch to using fopen().  This causes a warning;
I think a future change will be to define _CRT_SECURE_NO_WARNINGS
and use standard C library functions on Windows as on other
platforms.

Usd_CrateData is now careful to pass any path through TfAbsPath()
before passing it on to Usd_CrateDataImpl or CrateFile.  Without
this we can end up comparing paths with forward slashes to paths
with backslashes.  Paths differing in only that respect should
compare equal on Windows but don't.  As a result Save() tries to
copy to temporary and save to that but that fails because the file
is, in fact, already open for read Windows will not allow it to be
opened again for write.

Finally ~Usd_CrateDataImpl() now destroys the CrateFile object
synchronously.  Previously we couldn't be sure when we'd be able
do anything to that file on Windows because the file was closed
asynchronously.  For example, we can't delete a file until the
file is closed, and just that scenario played out in the
testUsdFileFormats test.

(Internal change: 1757600)