Replace *_s versions of functions with standard C library functions.
On Windows we get a warning when using certain C library functions,
like strcpy() and fopen(), unless _CRT_SECURE_NO_WARNINGS is set.
We now set that macro and use the old C library functions in some
places.

Note particularly the fopen() call in usd/crateData.cpp.  _fopen_s()
does *not* have the same behavior (with respect to sharing) as
fopen() so it's not a drop in replacement.  See change 1757600.

(Internal change: 1757991)