PxrUsdReadPointInstancer polish #2.
If a prototype prim is not a referenced model prim, make sure we build it "high" enough in the hierarchy to preserve material bindings.

Leverage PxrUsdKatanaReadXformable to gather primvars from the UsdGeomPointInstancer. Set them on both the "usd point instancer" location and its "instance array" child. Make sure to use the proper scopes ("point", "primitive").

Compute more data (e.g. the stage) upfront and pass them to the helpers as args.

(Internal change: 1758066)