Refactor internal value clip structures
Internal structures used for value clips now store
layer stacks and prim paths instead of Pcp nodes
since that's all that's needed by value resolution.

Also, Usd_ResolvedClipInfo now stores the source
layer stack and prim path instead of relying on
Usd_ResolveClipInfo to return the source node.
This will make it easier to modify that function
to return multiple results, which is needed for
an upcoming change to allow multiple sets of
clips on a single prim.

(Internal change: 1759386)