Update clip stitch utilities to write to clip sets
The clip stitching utilities now write to the default
clip set in the clips dictionary instead of the deprecated
metadata fields. To ease transition, these utilities
will revert to its old behavior of writing the individual
metadata fields if the USD_AUTHOR_LEGACY_CLIPS env var
is set, which is the current default.

(Internal change: 1760049)