Fix change processing bug with relocates
When a non-inert prim spec has been added to a layer,
that prim spec may have other scene description that
Sdf won't report explicitly in its change notification;
Sdf just considers those changes part of the addition
of the non-inert prim spec.

Because of this, when Pcp sees a non-inert prim spec
being added, it needs to scan to see if that spec (or
any of its descendant prims) has relocates. If so, all
affected layer stacks must recompute their composed
relocation tables to ensure they remain up-to-date.

(Internal change: 1763831)