[Hd/Hdx] Changes to prim gathering
A performance bug was reported, which cascaded into identifying correctness and performance issues with current prim gathering routines.

The originally reported performance issue, was caused by two things.
1) Dirty Lists was currently ignore collection exclude lists, so was over invalidating.
2) Prims in delegates not explictly in a collection's include list were being added to the dirty list (again over invalidating).

Further investigated revealed several other problems with prim gathering:
- Gathering of Draw Items failed to gather prims in a delegate's path was prefixed by that of another delegate that matched an include path.
- All the different Gathering algorithms did a linear search to find the end of the range
- Frequent creating and sorting list of prim paths.

This change attempts to standardize Prim Gathering into a single place, with a new class HdPrimGather.  The code has 3 entry points:
Filter() - Return a list of rprims that match the include path, but do not match the exclude path.
PredicatedFilter() - Similar to filter but paths must also pass an additional predicate (callback) test.
Subtree - Return a list of rprims that are prefixed by the given path.

After investigating a few different algorithms a binary search approach seemed to give us the best overral performance, with it's major cost being the setup cost of sorting the prim list.

Dirty List and Draw Items are the main consumer, with predicates to also filter the paths based on render tags and dirtiness (dirty list only).
Originally the predicate test was going to be multi-threaded, but performance results indicated that the time taken for the predicate test loop was not large enough for the multi-threaded overhead.

An additional consumer is Light gathering in the Light and Shadows tasks.

The Render Index' subtree interfaces, now use the gather's subtree function.

Changing to the new gather algorithm has also lead to the remove of the per delegate prim tracking in the render index.

The render index has also switched from storing the list of paths from a Set to a Vector.  An insertion sort is used to keep the vector sorted.

Changes:
- Added new HdPrimGather class implementing the gather algorithm.
- Removed HdDirtyList gather code and switched to the new class.
- HdDirtyList::ApplyEdit() will now do a full dirty list rebuild if there are exclude paths.
- Hd_PrimTypeIndex<PrimType>::InsertPrim() now does an insertion sort into the prim id vector (previously was a set).
- Hd_PrimTypeIndex<PrimType>::GetPrimSubtree() now uses HdPrimGather to do the subtree gather (was linear search to find the upper bound).
- Removed GetDelegateRprimIDs() and GetDelegateIDsWithDirtyRprims() from render index (now unused)
- Rprim ids now stored as a vector not a set in render index.
- Removed the delegate rprim map from the render index.
- AppendDrawItems() now always runs in parrellel and works on a per-thread list, rather than needing to make thread safe insertions into the list for every prim.
- Removed HdxSimpleLightTask::ComputeIncludedLights(), HdPrimGather::Filter() now used instead.

(Internal change: 1764747)
(Internal change: 1764935)