[Hd] Add Ext Computations to Hydra.
Ext Computations provide a mechanism for where a client can procedural generate a PrimVar.

This is experimental functionality, so right now is very limited in scope with many areas still left to do.
The purpose of this checkin is to get some exposure and feedback before fully implementing the feature.

Ext Computations are a new type of object stored in the render index, with the usual change tracking and syncing.

They are defined by a set of inputs and outputs.  An input to a computation may be bound to a scene input, which is obtained
by name from the scene delegate.  Alternatively an input may be bound to an output from another computation.

The output from a computation is a set of parallel arrays which may have differing types, but the same number of elements (i.e. SOA form).

Ext Computations use the resource registry system, to handle dependencies.  When it comes time to resolve an Ext computation,
the scene delegate is called to invoke the computation, using the computation context.

To use the result of a computation, an Rprim needs to define an ExtComputation PrimVar.  This is done by obtaining a list of
 ExtComputation PrimVars from the scene delegate, with an additional scene delegate function for obtaining the binding
information for that primVar.

Changes:
- Added ExtComputation object tracking to change tracker.
- Added new DirtyComputationPrimvarDesc dirty bit for rprims.
- ExtComputation object tracking and syncing  to render index.
- Expanded Scene Delegate API with new entry points for ExtComputations and ExtComputation PrimVars.
- Added a new RESOLVED_ERROR state to a HdBufferSource, meaning the buffer was resolved, but its contents is invalid.
- Added new HdBufferSource classes to represent computation input binding, computation processing and ExtComputation PrimVars.
- Added support to Basis curves to support ExtComputation PrimVars with Vertex Interpolation mode.

(Internal change: 1765350)