[Hd] Enforce Prim Ids have Delegate Ids as prefix. Hydra uses SdfPath's HasPrefix() test to determine if a prim is owned by a specific scene delegate.
This change adds a check that this condition is true, when the prim is added to the render index.

As the prior change ensures delegate ids are absolute paths, this change also has the side effect of also ensuring that Prim Ids are also absolute paths.

The Hd unit test delegate and testHdxCameraAndLight test were not following this rule, so they have been fixed up.

(Internal change: 1765868)