[Hd] Fix for incorrect and canonical trace scopes in SyncAll.
The trace scope for "Sync Ext Computations" was incorrectly defined at the scope of the HdRenderIndex::SyncAll() function. So it was including other scopes in this function as children and was not measuring the intended operation.

In Addition Trace scopes in SyncAll() were not in our canonical form.  Therefore, a new new macro has been added to Hf to do this automatically.

Changes:

-  Wrapped ext computations in {} to fix the scope.

- Added HF_TRACE_FUNCTION_SCOPE() macro which forms a tag with the "prettier" function name (used for function tags) followed by the
  scope tag in brackets "()"

(Internal change: 1769233)