[Hd] Add deferred sorting of prim ids. Internal performance tests indicated that a regression that was caused by initial population of a scene into a render index.  This is due to the insertion sort done while adding ids to the list of ids used by the prim gather algorithm.
This wasn't an unexpected regression, but wanted to get some scale of those numbers.  The best case scenario, was that the time taken from doing the sort every frame was moved to the initial population. Unfortunately, they were a lot larger.

To solve this problem, prim id sorting is deferred to when the id set is read.  To achieve this a new classes has been added to encapsulate the functionality.  It also uses a heuristic to determine the type of sorting to do, based on % of items that are new in the id set.

Changes:
- Added new Hd_SortedIds class that manages the list of id and performs the deferred sort.
- Switched R, S and B prims over to using the new class.
- Due to the defered sort, some methods can no longer be const.
- Added test for new class

(Internal change: 1769755)
(Internal change: 1769759)