[UsdImaging] Fix for increased memory usage caused by dirty bit propagation change The increase in memory usage was due to Subdiv tags in the UsdImaging delegate.
Prior to the dirty bit propagation change, when topology changes, we didn't tell the UsdImaging Delegate that we may pull on the subdiv tags.  So it didn't add the subdiv tags to the value cache when the delegate was synced.  Hydra would later pull on the value, when syncing the rprim, which resulted in a value cache miss in the  UsdImaging Delegate, so it would then need to go out to disk to get the value.

The dirty bit propagation change, changed the timing of when the subdiv tags dirty bit got set, so now when Hd syncs the delegate the subdiv tags bit is set when topology changes, so the delegate is adding the value to the value cache.

The cause of the memory difference, is that Hydra was only pulling the subdiv tags if the Rprim was refined, whereas with the dirty bit propagation change, the subdiv tag was always being loaded.

This change fixes this issue by adding logic to the usd imaging delegate to only process the subdiv tags dirty bit, if the prim is refined.  This has an advantage that it also removes the initial store of the subdiv tags in the value cache, because the dirty bit was set initally.

Changes:
- Added new IsRefined() method to usd imaging delegate.  This takes a usdPath (whereas GetRefineLevel is a interface function which works on render index path).

- Mesh Adapter only adds subdiv tags to the value cache when IsRefined is true.

(Internal change: 1770017)