Documentation ONLY enhancements/fixes for UsdGeom:
* Doc fix for UsdGeomPointBased and UsdGeomPointInstancer indicating the
units for velocities are distance / second, since that is what most DCC's
believe. Also add section in overview that describes how velocities are applied
to the computation of positions.

* Strengthen primvar doc on how to use indexed primvars

* Provide more explicit descriptions of matrix math and Xformable
xformOpOrder, and a full example.  Also moved linear algebra section to overview.

* Replace now-broken (due to prman doc churn) link to orientation with our
own version of the documentation in the overview, in a new section
"Winding Order, Orientation, and Surface Normals", which
also discusses using the matrix Determinant to determine when normals
need to be flipped.

* A few other small doc fixes.

(Internal change: 1770504)