Explicitly specify default value for GfHalf
GfHalf (really ilmbase's half) is a class type that holds
on to an unsigned short. Its default constructor does not
explicitly initialize this member, which caused the schema
to use an object with indeterminate value as the default
for half values in scene description. We now explicitly
specify a fallback of 0.0.

(Internal change: 1773378)