Fix validation mode in usdGenSchema
usdGenSchema's validation mode (which checks if the
code it generates differs from preexisting code) would
only run if there was a problem overwriting the
preexisting code. This meant the validation code
wasn't doing anything most of the time.

The validation code now runs earlier. In this mode,
usdGenSchema should never actually try to overwrite
anything: if the preexisting code is the same, there's
nothing to write, and if it's different, it should
report the diffs and exit immediately.

(Internal change: 1773858)