Make use of existing codeGen directives to allow classes like UsdGeomModelAPI to extend existing classes like UsdModelAPI through the facilities of codegen.  The sticking point had been that codeGen cannot handle two classes with the same prim name in two different, related, schema.usda files.  But UsdGeomModelAPI can have a prim name "GeomModelAPI" that inherits from core usd's "ModelAPI", while using the "className" customData to reset the name schemaGen uses for unprefixed codegen to "ModelAPI".
In a next step, we should make "extentsHint" be a core property of the
schema.

In a further future cleanup, we should make schemaGen refrain from populating
generatedSchema.usda and plugInfo.json with flattened "API" classes,
as the data for them is useless to the schema registry, and just
pollutes it.

(Internal change: 1774008)