Ignore arcs without opinions in instancing
Arcs without opinions no longer contribute to a prim's
instancing key. This allows for more sharing in common
scenarios involving inherit or specializes arcs where
the prim that is inherited or specialized is not
overridden in referencing layer stacks.

For example, given the following scenegraph:

Set
  SetA    -ref-> SetA
    Model          Model -ref-> Model -inh-> Class
  SetB    -ref-> SetB      |
    Model          Model --'

The prims /Set/SetA/Model and /Set/SetB/Model would
not share the same master, even though they ultimately
reference the exact same Model asset. This was because
there could have been different overrides for the
Class prim in the referenced SetA and SetB, which
would have caused these prims to compose differently.

With this change, Pcp will recognize that this isn't
the case -- there are no overrides for the class
in SetA or SetB -- so these prims can share the same
master.

(Internal change: 1774366)