[Hd] Remove trace scope from _AllocPrimId.
The Rprim insertion routine is very small routine, but is called at very high frequency when doing initial scene populations.  Recent performance analysis of the scalability of Rprim insertion showed that the function overhead was significant proportion of the time.  In particular, ~30% of the time was being spent in the performance tracing system.  There were actually two trace scope during Rprim insert, with the second inner one being for the _AllocPrimId method.
Changes:
- Removed the function trace scope for _AllocPrimId.

(Internal change: 1774793)