Allow primvars to contain arbitrarily nested sub-namespaces, such as "primvars:skel:jointIndices" or "primvars:very:deeply:nested:ns:mypv".
The behavior of UsdGeomImageable::GetPrimvar(TfToken const &name) is
to prepend "primvars:" onto 'name' if it is not already a prefix,
and return the result, which means we do not have any ambiguity between
the primvars "primvars:nsA:foo" and "primvars:nsB:foo".

Added UsdGeomPrimvar::GetPrimvarName() that simply strips the primvars
namespace, as GetBaseName() is no longer the appropriate name for
clients to use.

Also removes 'custom' param from Primvar ctors/creators, which was already
defaulted to false.

Updates client code that was using UsdGeomPrimvar::GetBaseName() to
use GetPrimvarName(), and, where approriate,
UsdGeomPrimvar::NameContainsNamespaces()

(Internal change: 1774878)
(Internal change: 1775066)