[UsdImaging] Removed Value Cache Allocation from UpdateForTimePrep. Recent performance data indicated significant performance gains could be had from removing value cache pre-allocation from UpdateForTimePrep.
The pre-allocation mechanism was a result of previous iterations of the value cache.  Since then, we've increased and improved threading throughout the system.  This includes the use of concurrent containers in the value cache.  As a result, the cost of the single-threaded allocation pass is
now a bottleneck rather than a help.

This change removes the value cache allocations from UpdateForTimePrep in all adapters.

Changes:
- Added empty implementation UpdateForTimePrep to base calls, so it is no longer a required function.
- Removed all value cache pre-allocations from UpdateForTimePrep in all prim adapters.
- Where possible removed empty UpdateForTimePrep from those adapters that no longer perform any operation during this step.

Notes:

Some adapters still use UpdateForTimePrep() to handle some inter-prim or inter-adapter dependencies.  This is undesirable behavior and we will be looking at ways to remove these, such that we can remove the UpdateForTimePrep step.

A similar change can be made to TrackVariabilityPrep()

(Internal change: 1775588)