usdGenSchema now uses installed schema.usda files
Libraries now install their schema.usda files as
resource files. Sublayer asset paths in these files
now use search paths instead of relative paths to
other schema.usda files in the source tree.
usdGenSchema has been modified to set up the
ArDefaultResolver to ensure these paths are
resolved to the correct location in the install.

This allows developers to run usdGenSchema on
schemas that sublayer in other schemas without
having the full USD source tree available. It also
simplifies the sublayer paths authored in a schema
which makes things easier to read.

This change means that opening a schema.usda
on a UsdStage no longer works without setting up
ArDefaultResolver with the necessary search path.
We think this is not a common workflow; if needed in
the future, we could provide a function to make this
easier.

Fixes #158

(Internal change: 1776071)