[UsdImaging] Cleanup of UsdImaging trying to update dirty bits during delegate sync. UsdImaging's UpdateForTime step, attempts to track and allows adapters to possibly change dirtybits.  The step is executed when Hydra attempts to sync the delegate.
The delegate takes this generated information and either tries to update the change tracker or return it to Hydra.  The return path to Hydra was removed in a previous iteration and updating the change tracker is incorrect as Hydra has already read the dirty bits and is working off local cached copies.
So the "results" dirty bits is a dead code path and such is being removed.

Changes:
- Removed Dirty Bits from UsdImagingDelegate's Worker.
- Removed ProcessUpdateResults() from UsdImagingDelegate's Worker.
- ExecuteWorkForTimeUpdate() no longer takes a boolean "UpdateDelegates" flag.
- Removed resultBits arg from UsdImagingPrimAdapter's UpdateForTime() method.
- Instancer Adapters no longer need to add in DirtyVisibility to either requestedBits and resultBits.  Removed always true if.

(Internal change: 1776327)