[UsdImaging] Cleanup of Prim Adapter's TrackVariability methods. UsdImaging's Prim Adapter's TrackVariability is used to determine which change bits are time varying and should be set if the time changes.
This code follows the same pattern as UpdateForTime(), with a single threaded TrackVariabilityPrep() pass followed by the multi-threaded TrackVariability() pass.

For most adapters, TrackVariabilityPrep() was a no-op.  For those, doing something, it was pre-allocating the value cache, which performance results indicated was not a help, so TrackVariabilityPrep() implementation has been removed from all adapters.

The TrackVariability() method took two HdDirtyBits paramters, the in parameter requestedBits and the out paramter dirtyBits.  This out parameter was badly named and was causing confusion, so has been renamed timeVaryingBits.

The in parameter requestedBits, was never needs to be set to anything other than "all bits".  This means the code was performing lots of if checks against requestedBits which were always going to be true.  So this parameter has been removed as well as all the if checking.

This means that UsdImagingDelegate::_Worker() could also be cleaned up to remove it's hierarchy of request bits, where they could come from the global state, task or "dirty bit" (which is also badly named) map.

Changes:
- Removed TrackVariabilityPrep() implementation from all adapters.
- Removed requestedBits param from TrackVariability().
- Renamed dirtyBits param to timeVaryingBits in TrackVariability().
- Removed _requestBits from  UsdImagingDelegate::_Worker.
- Removed default value from requestBits param of UsdImagingDelegate::_Worker's AddTask() method.

(Internal change: 1776638)