SdfListOp: Change addedOrExplicitItems from a mutable container proxy to a read-only GetAddedOrExplicitItems() method.
This is preparation for an upcoming change that will add
append and prepend capability to SdfListOp, at which point the
"addedOrExplicitItems" attribute that dynamically reflects either
the added list or the explicit list will no longer make sense --
since there will be more than one way for an item to be "added".

Instead, we define GetAddedOrExplicitItems() as computing
the result of applying the list-op to an empty list -- i.e. the
set of items the listop will cause to be added -- and return
that by value.  We also now wrap ApplyOperations().

(Internal change: 1776971)