Sdf: SdfListOp API now supports 'append' and 'prepend' operations.
summary

- SdfListOp now has append/prepend fields, and composes with them
- File i/o support: sdf, usda, usdc, mevnva
- Introduce crate version 0.2.0 to support the new ops
- USD_AUTHOR_APPEND_AS_ADD is an env var that controls whether
  Usd "Append()" API starts using the new append operators.
  Currently this is off.

Our rollout strategy is to get this change, representing
read and compose capability, out into the world and follow
up with a later change in a subsequent release to switch
over to authoring.

discussion

A common scenario is to add an arc in a layer, such as a
reference to a model.  Usually you expect a layer to introduce
opinions stronger than weaker layers, but list-op based
fields like composition arcs have never worked this way.
The reason is that those lists are ordered strong-to-weak
and the "add" operator will append an item to the end
(if it is not already present).  You can work around this
somewhat by also using a "reorder", but doing that succesfully
requires knowing all the other arcs in the list, so the
combination is both awkward and fragile.

To address this we simply add a "prepend" operator, which gives
a direct way to express adding a strongest entry to a list.
We also add "append" for symmetry.  These are both different
from "add" in that they also move existing entries to the
requested position.  A side benefit of this choice is that
if you restrict list-ops to only prepend/append/remove/explicit,
they become algebraically composable: you can put A over B
and get a list-op C that has the exact same effect.

Providing robust, simple flattening free of complex case
analysis is a long-term strategic goal of ours, and this
is a step towards that.

(Internal change: 1777058)
(Internal change: 1777109)