Fix bug in updating relocates for namespace edits
Pcp's namespace editing code was incorrectly using
lower_bound to determine if a path was prefixed by
any key in a map. This would only work if the path
happened to exactly match a key in the map. This
caused relocates to not be updated properly on
namespace edits to descendants of relocated prims.
We now use SdfPathFindLongestPrefix instead to
fix this issue.

(Internal change: 1777123)