Fix sublayer change processing bug with search paths
When handling the removal of a sublayer, Pcp would
use SdfLayer::FindRelativeToLayer to try to get a handle
to the sublayer being removed. This method currently
just anchors the sublayer path to the parent layer's
path, which would fail if this sublayer path was
actually a search path.

Pcp now uses SdfComputeAssetPathRelativeToLayer to
compute the appropriate asset path to find the sublayer.
In the future, SdfLayer::FindRelativeToLayer ought to
do the same thing, but that's a wider-reaching change.

(Internal change: 1777412)