Better handling for unexpected scene description
Issue warnings instead of coding errors for unexpected
scene description values. Coding errors are meant to
indicate a problem in the code, not in the external data
that code is reading. In addition, coding errors cause
exceptions in Python which can foul up consumers
that aren't handling them properly.

Also explicitly specify a fallback value in these cases.

(Internal change: 1777854)