USD: Change arc API methods from Append...() to Add...(), and add an optional 'position' argument.
Most of this change is mechanical API updates; the real
work is in lib/usd.

For rollout staging, the default value of position is
UsdListPositionTempDefault, which means to consult the
env setting of USD_AUTHOR_OLD_STYLE_ADD.  When true,
these API's continue to author "add" statements; when false,
they will use "prepend".  To avoid any change in behavior
with this change, the current env setting default is "true".

A subsequent changes will switch this default to "false",
actually changing our arc authoring behavior (but allowing
us to quickly switch back if needed).  A later change in
a future release will remove this env setting and position
value entirely, and change the default arg to
UsdListPositionFront, which authors a prepend.

(Internal change: 1781552)