[UsdImaging] Move UsdImaging to a common Prim Tracking and Updating mechanism (part 1).
As we are adding more prim types to the system, we wanted to make Scene Delegate more generic, relying on the Prim Adapters in UsdImaging to do the prim type specific work.

In this first part of the change, the focus is creating the common prim tracking mechanism in the delegate and fixing a few bugs along the way.  A 2nd part will add delegating responsibility to the adapters, with further changes migrating existing managed sprim types to the new mechanism.

Prior to this change, the adapter, usd prim and "dirty bits" for a prim were stored in separate look up maps.  Typical operations will require all this information, so this change consolidates all this information into a single "PrimInfo" structure.
This structure is indexed by "CachePath" (which is the Hydra Prim id minus the delegate id prefix).  This structure basically replaces the path adapter map and dirty bit map.

The dirty bits within the dirty map were used for two purposes: to indicate which bits should be marked as dirty when the time changes (the results of TrackVariability) and to mark bits that should be dirty for a single frame.  This was a bug, because once bits were marked as dirty once, they were never cleared.  Now in prim info the time varying and dirty state is tracked separately, and dirty state iscleared after its been updated.

This change also cleans up some of the dirty bit flow, which led to the bug with the time varying and dirty state being stored in the same place.

Changes:
- Renamed _PrepareWorkerForTimeUpdate() to _ApplyTimeVaryingState() as this function's purpose is to set dirty bits, not invoke the worker.
- Added _primInfo map, replacing _pathAdapterMap and _dirtyMap.
- Added GetPrimInfo() replacing  _AdapterLookupByPath() and _GetDirtyBits().
- Added _rprimPaths to identify a prim as an rprim.  This is a temporary measure until the functionality can be delegated to the adapters (this intended in part 2 of the change).
- MarkRprimOrInstancerDirty() no longer takes a cacheDirtyFlags.  Instead dirty bits in the prim info structure are always updated and the time varying bits left untouched.
- Added AddPrimInfo() and RemovePrimInfo() to UsdImagingIndexProxy, replacing AddDependency() and RemoveDependency().
- Renamed _depsToRemove to _primInfoToRemove.
- Switched UsdImagingIndexProxy's Insert*() method to taking a UsdPrim instead of a path, the cache path is obtained from the prim (and the adapters have the prim already, so no need to look it up).  The exception to this is InsertInstancer(), where the cache path and prim path may differ, so this method take the cache path as well as the usdPrim.
- UsdImagingDelegate::_Worker() no longer supports per-task dirty bits (These are obtained from the _PrimInfo structure).
- The worker's UpdateForTime() now cleans the prim's dirty bits.
- The delegate's Sync function now takes the dirty bits from hydra for rprims and merges them with it's own dirty state tracking to determine the dirty state of a prim.  (Note: Hydra still controls which rprims are dirty).

(Internal change: 1784001)
(Internal change: 1784065)